requirements.txt
astroid==2.4.2
awscli==1.18.170
awsebcli==3.19.2
boto3==1.16.10
botocore==1.19.10
cement==2.8.2
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
colorama==0.4.3
docutils==0.15.2
Flask==1.1.2
future==0.16.0
idna==2.10
isort==5.6.4
itsdangerous==1.1.0
Jinja2==2.11.2
jmespath==0.10.0
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
numpy==1.18.0
opencv-python==4.4.0.40
pathspec==0.5.9
Pillow==8.0.1
pyasn1==0.4.8
pylint==2.6.0
pypiwin32==223
python-dateutil==2.8.1
python-dotenv==0.15.0
pywin32==228
PyYAML==5.3.1
requests==2.24.0
rsa==4.5
s3transfer==0.3.3
semantic-version==2.5.0
six==1.15.0
termcolor==1.1.0
toml==0.10.2
urllib3==1.25.11
wcwidth==0.1.9
Werkzeug==1.0.1
wrapt==1.12.1